Convert trusted cluster to tctl handler format
