+++ tool/tctl/common/resource_command.go
@@ -111,7 +111,6 @@ Same as above, but using JSON output:
 // Initialize allows ResourceCommand to plug itself into the CLI parser
 func (rc *ResourceCommand) Initialize(app *kingpin.Application, _ *tctlcfg.GlobalCLIFlags, config *servicecfg.Config) {
 	rc.CreateHandlers = map[string]ResourceCreateHandler{
-		types.KindTrustedCluster:      rc.createTrustedCluster,
 		types.KindNetworkRestrictions: rc.createNetworkRestrictions,
 		types.KindDevice:              rc.createDevice,
 		types.KindOktaImportRule:      rc.createOktaImportRule,
@@ -450,40 +449,6 @@ func (rc *ResourceCommand) Create(ctx context.Context, client *authclient.Client
 	}
 }
 
-// createTrustedCluster implements `tctl create cluster.yaml` command
-func (rc *ResourceCommand) createTrustedCluster(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
-	tc, err := services.UnmarshalTrustedCluster(raw.Raw, services.DisallowUnknown())
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	// check if such cluster already exists:
-	name := tc.GetName()
-	_, err = client.GetTrustedCluster(ctx, name)
-	if err != nil && !trace.IsNotFound(err) {
-		return trace.Wrap(err)
-	}
-
-	exists := (err == nil)
-	if !rc.force && exists {
-		return trace.AlreadyExists("trusted cluster %q already exists", name)
-	}
-
-	//nolint:staticcheck // SA1019. UpsertTrustedCluster is deprecated but will
-	// continue being supported for tctl clients.
-	// TODO(bernardjkim) consider using UpsertTrustedClusterV2 in VX.0.0
-	out, err := client.UpsertTrustedCluster(ctx, tc)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	if out.GetName() != tc.GetName() {
-		fmt.Printf("WARNING: trusted cluster resource %q has been renamed to match root cluster name %q. this will become an error in future teleport versions, please update your configuration to use the correct name.\n", name, out.GetName())
-	}
-	fmt.Printf("trusted cluster %q has been %v\n", out.GetName(), UpsertVerb(exists, rc.force))
-	return nil
-}
-
 // createNetworkRestrictions implements `tctl create net_restrict.yaml` command.
 func (rc *ResourceCommand) createNetworkRestrictions(ctx context.Context, client *authclient.Client, raw services.UnknownResource) error {
 	newNetRestricts, err := services.UnmarshalNetworkRestrictions(raw.Raw, services.DisallowUnknown())
@@ -723,11 +688,6 @@ func (rc *ResourceCommand) Delete(ctx context.Context, client *authclient.Client
 			return trace.Wrap(err)
 		}
 		fmt.Printf("reverse tunnel %v has been deleted\n", ref.Name)
-	case types.KindTrustedCluster:
-		if err = client.DeleteTrustedCluster(ctx, ref.Name); err != nil {
-			return trace.Wrap(err)
-		}
-		fmt.Printf("trusted cluster %q has been deleted\n", ref.Name)
 	case types.KindRemoteCluster:
 		if err = client.DeleteRemoteCluster(ctx, ref.Name); err != nil {
 			return trace.Wrap(err)
@@ -976,25 +936,6 @@ func (rc *ResourceCommand) getCollectionByRef(ctx context.Context, client *authc
 		}
 
 		return &reverseTunnelCollection{tunnels: tunnels}, nil
-	case types.KindTrustedCluster:
-		if ref.Name == "" {
-			// TODO(okraport): DELETE IN v21.0.0, replace with regular Collect
-			trustedClusters, err := clientutils.CollectWithFallback(
-				ctx,
-				client.ListTrustedClusters,
-				client.GetTrustedClusters,
-			)
-			if err != nil {
-				return nil, trace.Wrap(err)
-			}
-
-			return &trustedClusterCollection{trustedClusters: trustedClusters}, nil
-		}
-		trustedCluster, err := client.GetTrustedCluster(ctx, ref.Name)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		return &trustedClusterCollection{trustedClusters: []types.TrustedCluster{trustedCluster}}, nil
 	case types.KindRemoteCluster:
 		if ref.Name == "" {
 			remoteClusters, err := client.GetRemoteClusters(ctx)
